mask root dir in compilation options widget
